requirements.txt
accelerate
datasets
matplotlib
numpy
openai
pillow
protobuf
scikit-learn
sentencepiece
timm
torch
torchvision
tqdm
transformers
vllm
